Move summary cards out of debug modal and into Workbook tab
The Sheets / File Size / Student History Comments cards now live
under the Workbook tab's main section in Settings rather than
inside the Workbook Debug modal. Settings owns the load logic
(loadWorkbookSummary + loadFileSize) and refreshes whenever the
Workbook tab's main section becomes active. The Workbook Debug
modal is back to its tree/json view only.
